+++ driver/web/adapter.go
@@ -105,7 +105,9 @@ func (we Adapter) Start() {
 	adminSubrouter.HandleFunc("/applications/{id}", we.wrapFunc(we.adminApisHandler.getApplication, we.auth.adminAuth)).Methods("GET")
 	adminSubrouter.HandleFunc("/applications", we.wrapFunc(we.adminApisHandler.getApplications, we.auth.adminAuth)).Methods("GET")
 
-	adminSubrouter.HandleFunc("/application-permissions", we.wrapFunc(we.adminApisHandler.createApplicationPermission, we.auth.adminAuth)).Methods("POST")
+	adminSubrouter.HandleFunc("/permissions", we.wrapFunc(we.adminApisHandler.createPermission, we.auth.adminAuth)).Methods("POST")
+	adminSubrouter.HandleFunc("/permissions", we.wrapFunc(we.adminApisHandler.updatePermission, we.auth.adminAuth)).Methods("PUT")
+
 	adminSubrouter.HandleFunc("/application-roles", we.wrapFunc(we.adminApisHandler.createApplicationRole, we.auth.adminAuth)).Methods("POST")
 	adminSubrouter.HandleFunc("/application-api-keys", we.wrapFunc(we.adminApisHandler.getApplicationAPIKeys, we.auth.adminAuth)).Methods("GET")