+++ driver/web/apis_admin.go
@@ -614,22 +614,48 @@ func (h AdminApisHandler) getApplications(l *logs.Log, r *http.Request, claims *
 	return l.HttpResponseSuccessJSON(data)
 }
 
-//createApplicationPermission creates an application permission
-func (h AdminApisHandler) createApplicationPermission(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+//createPermission creates an permission
+func (h AdminApisHandler) createPermission(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
 	data, err := ioutil.ReadAll(r.Body)
 	if err != nil {
 		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
 	}
 
-	var requestData Def.ReqApplicationPermissionsRequest
+	var requestData Def.ReqPermissionsRequest
 	err = json.Unmarshal(data, &requestData)
 	if err != nil {
-		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypeApplicationPermission, nil, err, http.StatusBadRequest, true)
+		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypePermission, nil, err, http.StatusBadRequest, true)
 	}
 
-	_, err = h.coreAPIs.Administration.AdmCreateApplicationPermission(requestData.Name, requestData.AppId)
+	if requestData.ServiceIds != nil {
+		_, err = h.coreAPIs.Administration.AdmCreatePermission(requestData.Name, *requestData.ServiceIds)
+	} else {
+		_, err = h.coreAPIs.Administration.AdmCreatePermission(requestData.Name, nil)
+	}
+
 	if err != nil {
-		return l.HttpResponseErrorAction(logutils.ActionGet, model.TypeApplicationPermission, nil, err, http.StatusInternalServerError, true)
+		return l.HttpResponseErrorAction(logutils.ActionCreate, model.TypePermission, nil, err, http.StatusInternalServerError, true)
+	}
+
+	return l.HttpResponseSuccess()
+}
+
+//updatePermission updates an permission
+func (h AdminApisHandler) updatePermission(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+	data, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
+	}
+
+	var requestData Def.ReqPermissionsRequest
+	err = json.Unmarshal(data, &requestData)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypePermission, nil, err, http.StatusBadRequest, true)
+	}
+
+	_, err = h.coreAPIs.Administration.AdmUpdatePermission(requestData.Name, requestData.ServiceIds)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUpdate, model.TypePermission, nil, err, http.StatusInternalServerError, true)
 	}
 
 	return l.HttpResponseSuccess()
@@ -666,12 +692,12 @@ func (h AdminApisHandler) grantAccountPermissions(l *logs.Log, r *http.Request, 
 	var requestData Def.ReqAccountPermissionsRequest
 	err = json.Unmarshal(data, &requestData)
 	if err != nil {
-		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypeApplicationPermission, nil, err, http.StatusBadRequest, true)
+		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypePermission, nil, err, http.StatusBadRequest, true)
 	}
 
-	err = h.coreAPIs.Administration.AdmGrantAccountPermissions(requestData.AccountId, requestData.AppId, requestData.Permissions)
+	err = h.coreAPIs.Administration.AdmGrantAccountPermissions(requestData.AccountId, requestData.Permissions)
 	if err != nil {
-		return l.HttpResponseErrorAction(actionGrant, model.TypeApplicationPermission, nil, err, http.StatusInternalServerError, true)
+		return l.HttpResponseErrorAction(actionGrant, model.TypePermission, nil, err, http.StatusInternalServerError, true)
 	}
 
 	return l.HttpResponseSuccess()